requirements.txt
pdfplumber
camelot-py
PyPDF2
PyMuPDF
opencv-python 
pytesseract